requirements.txt
autopep8==2.0.*
flake8==6.0.*
python-dotenv==1.0.*
requests==2.28.*
